Move events into separate files
Also Nuimo Glyphs

This also organizes the imports
