+++ bonk/src/devices/nuimo/index.ts
@@ -1,1 +1,3 @@
-export { Nuimo, EVENTS, GLYPHS } from './device';
+export { Nuimo } from './device';
+export { NuimoEvents } from './events';
+export { NuimoGlyphs } from './glyphs';
