+++ bonk/src/devices/powermate/events.ts
@@ -1,0 +1,11 @@
+/** Events for PowerMate button and rotation */
+export const PowerMateEvents: { [eventName: string]: string } = {
+  SINGLE_PRESS: 'singlePress',
+  DOUBLE_PRESS: 'doublePress',
+  TRIPLE_PRESS: 'triplePress',
+  LONG_PRESS: 'longPress',
+  CLOCKWISE: 'clockwise',
+  COUNTERCLOCKWISE: 'counterclockwise',
+  PRESS_CLOCKWISE: 'pressClockwise',
+  PRESS_COUNTERCLOCKWISE: 'pressCounterclockwise',
+};
