+++ bonk/src/devices/powermate/index.ts
@@ -1,1 +1,2 @@
-export { PowerMate, EVENTS } from './device';
+export { PowerMate } from './device';
+export { PowerMateEvents } from './events';
