+++ bonk/src/devices/sonos/events.ts
@@ -1,0 +1,6 @@
+/** Events for Sonos play state updates */
+export const SonosEvents: { [eventName: string]: string } = {
+  PLAYING: 'isPlaying',
+  PAUSED: 'isPaused',
+  GROUP_CHANGED: 'groupChanged',
+};
