+++ bonk/src/index.ts
@@ -1,10 +1,6 @@
-import {
-  Nuimo,
-  EVENTS as NuimoEvents,
-  GLYPHS as NuimoGlyphs,
-} from './devices/nuimo';
-import { PowerMate, EVENTS as PowerMateEvents } from './devices/powermate';
-import { Sonos, EVENTS as SonosEvents } from './devices/sonos';
+import { Nuimo, NuimoEvents, NuimoGlyphs } from './devices/nuimo';
+import { PowerMate, PowerMateEvents } from './devices/powermate';
+import { Sonos, SonosEvents } from './devices/sonos';
 
 const powermate = new PowerMate();
 const sonos = new Sonos();
